Major design changes
Introduced a sort of Self-Decorating/Adapter design for the game objects
that are going to be drawn by the game engine, in order to decouple them
from the game engine itself.
Added IGameObject, IDrawableGameObject, IDrawableAttribute.
Added DrawableGameObject, DrawableChessBoard, DrawableChessPiece,
DrawableField.
Made changes to ChessBoard, ChessPiece and Field, in order to implement
the new interfraces.
Made changes to almost all existing interfaces, in order to implement
the updated design.
Current Unit tests are NOT up-to-date. Mocking objects are not required
anymore and have to be replaced with the actual objects from the
BattleChess.Infrastructure namespace.